+++ js/dataset/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/dataset
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [[`3c3cf86`](https://github.com/gradio-app/gradio/commit/3c3cf8618a8cad1ef66a7f96664923d2c9f5e0e2), [`3f139c7`](https://github.com/gradio-app/gradio/commit/3f139c7c995f749562bb007d2a567bb167669de9)]:
+  - @gradio/client@0.10.1
+  - @gradio/upload@0.6.1
+
 ## 0.1.16
 
 ### Patch Changes